# Conflicts:
#	src/main/java/ir/alzahra/offerBaz/control/IOfferService.java
#	src/main/java/ir/alzahra/offerBaz/facade/impl/OfferFacade.java
#	src/main/java/ir/alzahra/offerBaz/model/dao/IProductDao.java

+++ src/main/java/ir/alzahra/offerBaz/control/impl/OfferServiceImpl.java
@@ -1,13 +1,11 @@
 package ir.alzahra.offerBaz.control.impl;
 
 
+import ir.alzahra.offerBaz.control.IOfferCheckService;
 import ir.alzahra.offerBaz.control.IOfferService;
-import ir.alzahra.offerBaz.dto.BankDTO;
-import ir.alzahra.offerBaz.dto.searchParameter.ProductSearchParam;
 import ir.alzahra.offerBaz.exception.BaseException;
 import ir.alzahra.offerBaz.model.dao.IBankDao;
 import ir.alzahra.offerBaz.model.dao.IProductDao;
-import ir.alzahra.offerBaz.model.dao.impl.ProductDao;
 import ir.alzahra.offerBaz.model.entity.BankEntity;
 import ir.alzahra.offerBaz.model.entity.ProductEntity;
 import ir.alzahra.offerBaz.notify.CustomSpringEvent;
@@ -18,6 +16,7 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * @Author: Mobina Pak
@@ -34,14 +33,12 @@ public class OfferServiceImpl implements IOfferService {
 
     @Autowired
     private CustomSpringEvent applicationEventPublisher;
+
+    @Autowired
+    private IOfferCheckService offerCheckService;
+
     private String finalcode;
 
-    @Override
-    public void insertProduct(ProductEntity productEntity) throws BaseException{
-        productDao.insert(productEntity);
-        applicationEventPublisher.notify("product.insert.success", NotificationType.Info);
-
-    }
 
     public String generateUniqueCode(String name2) throws BaseException {
        long code = productDao.getCountOfRecord()+1000;
@@ -58,6 +55,12 @@ public class OfferServiceImpl implements IOfferService {
 
     @Override
     public void insertBank(BankEntity bank) throws BaseException {
+        offerCheckService.checkInsertBank(bank);
+       BankEntity duplicateBank= bankDao.findByName(bank.getName());
+       if (Objects.nonNull(duplicateBank))
+           throw new BaseException("bank.insert.duplicateBankName");
+       if (Objects.nonNull(bankDao.findBankByAbbreviation(bank.getNameAbbreviation())))
+           throw new BaseException("bank.insert.duplicateMokhafafName");
         bankDao.insert(bank);
         applicationEventPublisher.notify("bank.insert.success", NotificationType.Info);
 
@@ -65,37 +68,56 @@ public class OfferServiceImpl implements IOfferService {
 
     @Override
     public List<BankEntity> getAllBanks() throws BaseException {
-       return bankDao.getAllBanks();
+        List<BankEntity> banks=bankDao.getAllBanks();
+        if (Objects.isNull(banks)|| banks.size()==0)
+            throw new BaseException("bank.find.emptyBankList");
+       return banks;
     }
 
     @Override
-    public void updateBank(BankEntity bankEntity) throws BaseException {
+    public void updateBank(BankEntity bankEntity,String trackCode) throws BaseException {
         bankDao.update(bankEntity);
-        applicationEventPublisher.notify("product.insert.success",new Object[]{bankEntity.getProducts().get(0).getUniqueCode()}, NotificationType.Info);
+        applicationEventPublisher.notify("product.insert.success",new Object[]{trackCode}, NotificationType.Info);
 
     }
 
     @Override
     public BankEntity findBankByName(String bankName) throws BaseException {
-       return bankDao.findByName(bankName);
+        BankEntity bank=bankDao.findByName(bankName);
+        if(Objects.isNull(bank))
+            throw new BaseException("bank.find.notFoundBank");
+        else
+       return bank;
     }
 
     @Override
     public ProductEntity findProductByCode(String trackingCode) throws BaseException {
-        return productDao.findProductByCode(trackingCode);
+        ProductEntity p=productDao.findProductByCode(trackingCode);
+        if (Objects.isNull(p))
+            throw new BaseException("product.insert.notFoundProduct");
+        else return p;
     }
 
     @Override
     public void updateProduct(ProductEntity productEntity) throws BaseException {
+        if (Objects.isNull(productEntity.getId()))
+            throw new BaseException("product.edit.notFoundForEdit");
+        else{
+            offerCheckService.checkUpdateProduct(productEntity);
+            checkExistProduct(productEntity.getProductName());
         productDao.update(productEntity);
         applicationEventPublisher.notify("product.update.success", NotificationType.Info);
+        }
     }
 
     @Override
     public void deleteProduct(ProductEntity productEntity) throws BaseException {
-        productDao.delete(productEntity);
-        applicationEventPublisher.notify("product.delete.success", NotificationType.Info);
-
+        if (Objects.isNull(productEntity.getId()))
+            throw new BaseException("product.delete.notFoundForDelete");
+        else {
+            productDao.delete(productEntity);
+            applicationEventPublisher.notify("product.delete.success", NotificationType.Info);
+        }
     }
 
     @Override
@@ -111,7 +133,18 @@ public class OfferServiceImpl implements IOfferService {
 
     @Override
     public List<BankEntity> searchBankByParam(String bankName) throws BaseException {
-        return bankDao.findBankByParam(bankName);
+        List<BankEntity> bankEntity=bankDao.findBankByParam(bankName);
+        if(Objects.isNull(bankEntity) || bankEntity.size()==0)
+            throw new BaseException("bank.find.notFoundBank");
+        else return bankEntity;
+    }
+
+    @Override
+    public void checkExistProduct(String name) throws BaseException {
+        ProductEntity p =productDao.findByName(name);
+        if (Objects.nonNull(p))
+            throw new BaseException("product.insert.exist");
+
     }
 
     @Override
